Commenting out def for MBEDTLS_FS_IO
MBEDTLS_FS_IO is defined to enable functions that use the filesystem. For the scope of our project, we don't need this. I am commenting this out because allowing MBEDTLS_FS_IO makes the compiler find errors in an array that is defined with a size 0. This is done here because the array size depends on the user input.